+++ src/api/schema/response.rs
@@ -9,7 +9,7 @@ use super::integrations::{
 use super::panes::{
     LayoutDescription, PaneEdgesResult, PaneFocusDirectionResult, PaneInfo, PaneLayoutSnapshot,
     PaneMoveResult, PaneNeighborResult, PaneProcessInfo, PaneReadResult, PaneResizeResult,
-    PaneSwapResult, PaneZoomResult,
+    PaneSearchResult, PaneSwapResult, PaneZoomResult,
 };
 use super::plugins::{
     InstalledPluginInfo, PluginActionInfo, PluginCommandLogInfo, PluginInvocationContext,
@@ -175,6 +175,9 @@ pub enum ResponseResult {
     },
     PaneResize {
         resize: PaneResizeResult,
+    },
+    PaneSearch {
+        search: PaneSearchResult,
     },
     PaneRead {
         read: PaneReadResult,